@@ -3,16 +3,17 @@
 
 import React from "react";
 import styled from "styled-components";
+import { text } from "../Localisation/text";
 
 const A = styled.a`
   text-decoration: none;
   color: black;
 `;
 const Span = styled.span`
-  color: ${(props) => (props.dark ? "yellow" : "blue")};
+  color: ${(props) => (props.dark ? "gold" : "blue")};
   border-radius: 10px;
   &:hover {
-    background-color: ${(props) => (props.dark ? "purple" : "gold")};
+    background-color: ${(props) => (props.dark ? "blue" : "gold")};
   }
   &:active {
     background-color: yellowgreen;
@@ -22,11 +23,11 @@ const P = styled.p`
   margin: 4px;
 `;
 
-function Attribution({ darkMode }) {
+function Attribution({ darkMode, language }) {
   return (
     <footer className="attribution">
       <P>
-        Attribution:
+        {text[language].attributionText}
         <A href="https://sunrise-sunset.org/" target="_blank">
           <Span dark={darkMode ? true : false}> Sunrise Sunset,</Span>
         </A>
@@ -34,19 +35,18 @@ function Attribution({ darkMode }) {
           <Span dark={darkMode ? true : false}> Stormglass,</Span>
         </A>
         <A href="https://www.geoapify.com/" target="_blank">
-          <Span dark={darkMode ? true : false}> Geoapify,</Span>
-        </A>
-        <A href="https://www.mapbox.com/" target="_blank">
-          <Span dark={darkMode ? true : false}> Mapbox</Span>
+          <Span dark={darkMode ? true : false}> Geoapify</Span>
         </A>
       </P>
       <P>
-        <strong>Nature is unpredictable.</strong>
-        <em> Use the information presented at your own risk.</em>
+        <strong>{text[language].attributionWarning}</strong>
+        <em> {text[language].attributionReminder}</em>
       </P>
-      Created Oct 2021 by ZY of GA SEI-32. Github code
+      {text[language].attributionLink}
       <A href="https://github.com/zyteo/LighTide" target="_blank">
-        <Span dark={darkMode ? true : false}> here.</Span>
+        <Span dark={darkMode ? true : false}>
+          {text[language].attributionLinkSpan}
+        </Span>
       </A>
     </footer>
   );
