@@ -10,16 +10,17 @@ import {
   VictoryScatter,
   VictoryTheme,
 } from "victory";
+import { text } from "../Localisation/text";
 
-function Chart({ tide }) {
+function Chart({ tide, language }) {
   // merge both zoom and voronoi containers
   const VictoryZoomVoronoiContainer = createContainer("zoom", "voronoi");
   // create array to store the data for the time series chart
   const tideSeries = [];
   const tidedetails = tide?.data?.map((ele) => {
-    let timeTideDetails = new Date(ele.time);
+    let timeTideDetails = new Date(ele?.time);
     // add data for the time series chart
-    tideSeries.push({ x: timeTideDetails, y: ele.height });
+    tideSeries.push({ x: timeTideDetails, y: ele?.height });
     return null;
   });
 
@@ -27,12 +28,12 @@ function Chart({ tide }) {
     <div className="chart">
       <p>
         <strong>
-          <u>Tide Chart</u>
+          <u>{text[language].chartTitle}</u>
         </strong>
       </p>
 
       {/* ternary conditional operator here in case tide limit reached. */}
-      {tide.data ? (
+      {tide?.data ? (
         <>
           <div className="timeseries">
             <VictoryChart
@@ -42,15 +43,17 @@ function Chart({ tide }) {
               theme={VictoryTheme.material}
               containerComponent={
                 <VictoryZoomVoronoiContainer
-                  labels={({ datum }) => `${datum.x} \n ${datum.y} meters`}
+                  labels={({ datum }) =>
+                    `${datum.x} \n ${datum.y} ${text[language].tideMetres}`
+                  }
                 />
               }
             >
               <VictoryLabel
                 text={
-                  "Time Series (" +
+                  `${text[language].chartTimeSeries} (` +
                   `${tide?.meta?.station?.name}` +
-                  " station)"
+                  ` ${text[language].mapsStation})`
                 }
                 x={225}
                 y={30}
@@ -75,7 +78,7 @@ function Chart({ tide }) {
           </div>
         </>
       ) : (
-        <p>Tide request limit reached for today, so no chart generated!</p>
+        <p>{text[language].chartLimitReached1}</p>
       )}
     </div>
   );
