@@ -4,12 +4,12 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
+import { text } from "../Localisation/text";
 
 const Navbar = styled.nav`
   display: flex;
   align-items: center;
-  width: 100%;
-  height: 6vh;
+  width: 100vw;
   padding: 0 20px;
   background-color: lightskyblue;
   &:hover {
@@ -79,24 +79,40 @@ const CheckBox = styled.input`
   }
 `;
 
-function Nav({ handleDarkMode }) {
+const Select = styled.select`
+  margin-left: 20px;
+  font-size: 16px;
+`;
+
+const Option = styled.option`
+  font-size: 16px;
+`;
+function Nav({ handleDarkMode, language, setLanguage }) {
+  const handleLanguage = (event) => {
+    setLanguage(event.target.value);
+  };
   return (
     <>
       <Navbar>
         <LinkStyled to="/">
-          <Li>Home</Li>
+          <Li>{text[language].navHome}</Li>
         </LinkStyled>
         <LinkStyled to="/maps">
-          <Li>LighTide</Li>
+          <Li>{text[language].navLT}</Li>
         </LinkStyled>
         <LinkStyled to="/results">
-          <Li>Results</Li>
+          <Li>{text[language].navResults}</Li>
         </LinkStyled>
         <CheckBoxWrapper>
           <CheckBox id="checkbox" type="checkbox" onClick={handleDarkMode} />
           <CheckBoxLabel htmlFor="checkbox" />
         </CheckBoxWrapper>
         &#127769;
+        <Select onClick={(event) => handleLanguage(event)}>
+          <Option value="English">English</Option>
+          <Option value="简体中文">简体中文</Option>
+          <Option value="繁体中文">繁体中文</Option>
+        </Select>
       </Navbar>
     </>
   );
